Remove unused legacy classes
