+++ sceneview/src/main/java/io/github/sceneview/scene/Transformable.kt
@@ -1,0 +1,13 @@
+package io.github.sceneview.scene
+
+interface Transformable {
+    val isFocusable: Boolean
+    val editModes: Set<EditMode>
+
+    /**
+     * Allowed ways to manipulate the node.
+     */
+    enum class EditMode {
+        ROTATE, SCALE, MOVE
+    }
+}
